pkg: move temp_git_repo to scm/git
